Adjusting value property to double

+++ src/EcleticaBeerControl.Domain/Entities/Timeseries/Temperature.cs
@@ -4,6 +4,6 @@
     {
         public required DateTime Time { get; init; }
         public required string DeviceId { get; init; } = string.Empty;
-        public long? Value { get; init; }
+        public double? Value { get; init; }
     }
 }
