+++ src/EcleticaBeerControl.Worker/Payloads/Temperature.cs
@@ -4,7 +4,7 @@ namespace EcleticaBeerControl.Worker.Payloads
 {
     public sealed record Temperature
     {
-        public long Degrees { get; set; }
+        public double Degrees { get; set; }
         public DegreeFormat DegreeFormat { get; set; }
     }
 }
